Test Aws::Rails::Railtie
