+++ .gitignore
@@ -5,4 +5,6 @@
 coverage
 doc
 Gemfile.lock
+test/dummy/log/
+test/dummy/tmp/
 vendor
