+++ test/aws/rails/mailer_test.rb
@@ -1,6 +1,20 @@
 require 'test_helper'
 require 'mail'
 
+class TestMailer < ActionMailer::Base
+  layout nil
+
+  def deliverable(options = {})
+    mail(
+      body: options[:body],
+      delivery_method: :aws_sdk,
+      from: options[:from],
+      subject: options[:subject],
+      to: options[:to]
+    )
+  end
+end
+
 module Aws
   module Rails
     class MailerTest < Minitest::Test
