+++ test/dummy/app/assets/config/manifest.js
@@ -1,0 +1,1 @@
+// This file exists to appease Sprockets 4
